Add NormalizeString to DorchTypes
Add Medituj() for skill meditation
